+++ open-ig/src/hu/openig/music/OggMusic.java
@@ -80,7 +80,7 @@ public class OggMusic {
 			// catch(Exception ee){
 			// System.out.println(ee);
 			// }
-			AudioFormat audioFormat = new AudioFormat((float) rate, 16,
+			AudioFormat audioFormat = new AudioFormat(rate, 16,
 					channels, true, // PCM_Signed
 					false // littleEndian
 			);
